applying useReducer() instead of useState() to understand advantages and disadvantages.
If applied, this commit will have applied useReducer() instead of useState() to understand advantages and disadvantages.

useReducer() should definitely considered if we are dealing with many states or states that contain objects.
In a primary approach useState() should be the first option, but in case of more state complexity should be fine to use useReducer().